scrapers/requirements.txt
# Telecommunications Intelligence Scraper Requirements
# Core dependencies for web scraping and data processing

# HTTP client for API requests
aiohttp>=3.8.0

# Data processing and manipulation
pandas>=1.5.0
numpy>=1.21.0

# Date and time handling
python-dateutil>=2.8.0

# JSON handling and data serialization
ujson>=5.0.0

# Logging and monitoring
structlog>=22.0.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: For advanced text processing
# nltk>=3.7
# spacy>=3.4.0

# Optional: For database storage
# sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0

# Optional: For email notifications
# sendgrid>=6.9.0

# Optional: For web scraping (if needed beyond Serper API)
# playwright>=1.20.0
# beautifulsoup4>=4.11.0
# requests>=2.28.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0
